refactor deprecation tests
